* include/cln/object.h (cl_combine): define additional signatures, if           HAVE_LONGLONG is defined, in order to keep the compiler happy.         * src/base/cl_macros.h: include "cln/types.h", since we need HAVE_DD...         * src/base/cl_macros.h (bit): ...for this macro...         * src/base/cl_macros.h (minus_bit): ...and this one.         * src/base/cl_low.h: include "cln/types.h", since we need HAVE_DD...         * src/base/cl_low.h (logcount_64): ...for this macro.         * src/base/random/cl_UL_random.cc (random32): if HAVE_DD a is an ULL.         * src/integer/gcd/cl_I_gcd_aux2.cc (floorDD): fixed algorithmic bug           that turned up when intDsize==32 and cl_word_size==64.         * src/float/dfloat/elem/cl_DF_div.cc (operator/): fixed a missing cast           to uint64 that turned up when intDsize==32 and cl_word_size==64.

+++ src/float/dfloat/elem/cl_DF_div.cc
@@ -145,7 +145,7 @@ const cl_DF operator/ (const cl_DF& x1, const cl_DF& x2)
        // q = 2^32*manthi+mantlo.
        #if (cl_word_size==64)
        #if (intDsize<=32)
-       mantx = (manthi<<32) | (uint64)mantlo;
+       mantx = ((uint64)manthi<<32) | (uint64)mantlo;
        #endif
        if (mantx >= bit(DF_mant_len+2))
          // Quotient >=2^54 -> 2 Bits wegrunden
